requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.9.2
faiss-cpu==1.8.0.post1
numpy==1.26.4
scikit-learn==1.5.2
PyPDF2==3.0.1
google-generativeai==0.7.2
tenacity==8.5.0
streamlit==1.39.0
orjson==3.10.7

